Pillow error solved
After running this code , the user will be able to see a window flashing up and then the user have to click on the question box in order to see the question . Then the user can accordingly answer the questions in the answer box given for each question.
After the user is done with filling the answers in the given boxes , he/she MUST click on "SUBMIT' button and then must also close the window manually or shall close it by clicking on the "Exit Program button.

The user is done submitting the program , he/she can get the result printed up in the RUN shell/window of the code window.
The print will be in format of "Question" = "The answer which the user have written".

Questions topic = Country PM's , Indian States CM's , Capital of Countries  and Capital of Indian States.

The questions are randomly selected from a set of 34 questions , which i made.

NOTE :
This code does not checks whether the answer the user have given is accordingly right or wrong. 
This code only till now stores the question and answers given by the user temporarily.

This project constructed with the help of  Python language , Tkinter (GUI).